Move conversion files into a subdirectory
Trying to logically split the long list of dozens of metrics files.

Bug: 550013241
