+++ base/cvd/cuttlefish/host/libs/metrics/BUILD.bazel
@@ -16,7 +16,7 @@ cf_cc_library(
         "//cuttlefish/host/commands/cvd/instances",
         "//cuttlefish/host/libs/metrics:device_event_type",
         "//cuttlefish/host/libs/metrics:guest_metrics",
-        "//cuttlefish/host/libs/metrics:metrics_conversion",
+        "//cuttlefish/host/libs/metrics/conversion:metrics_data",
         "//cuttlefish/host/libs/metrics:metrics_orchestration",
         "//cuttlefish/host/libs/metrics:notification",
         "//cuttlefish/host/libs/metrics:parsed_flags",
@@ -43,23 +43,6 @@ cf_cc_library(
 )
 
 cf_cc_library(
-    name = "fetch_conversion",
-    srcs = ["fetch_conversion.cc"],
-    hdrs = ["fetch_conversion.h"],
-    deps = [
-        "//cuttlefish/host/commands/cvd/fetch:builds",
-        "//cuttlefish/host/libs/metrics:fetch_metrics",
-        "//cuttlefish/host/libs/web:android_build",
-        "//external_proto:cf_fetch_build_cc_proto",
-        "//external_proto:cf_fetch_builds_cc_proto",
-        "//external_proto:cf_fetch_complete_cc_proto",
-        "//external_proto:cf_fetch_failure_cc_proto",
-        "//external_proto:cf_fetch_start_cc_proto",
-        "//external_proto:cf_metrics_event_v2_cc_proto",
-    ],
-)
-
-cf_cc_library(
     name = "fetch_metrics",
     srcs = ["fetch_metrics.cc"],
     hdrs = ["fetch_metrics.h"],
@@ -81,7 +64,7 @@ cf_cc_library(
         "//cuttlefish/host/commands/cvd/fetch:fetch_cvd",
         "//cuttlefish/host/commands/cvd/fetch:fetch_cvd_parser",
         "//cuttlefish/host/libs/metrics:fetch_metrics",
-        "//cuttlefish/host/libs/metrics:metrics_conversion",
+        "//cuttlefish/host/libs/metrics/conversion:metrics_data",
         "//cuttlefish/host/libs/metrics:metrics_orchestration",
         "//cuttlefish/host/libs/metrics:notification",
         "//cuttlefish/result",
@@ -125,22 +108,6 @@ cf_cc_library(
 )
 
 cf_cc_library(
-    name = "guest_conversion",
-    srcs = ["guest_conversion.cc"],
-    hdrs = ["guest_conversion.h"],
-    deps = [
-        "//cuttlefish/host/libs/config:data_image_policy",
-        "//cuttlefish/host/libs/config:gpu_mode",
-        "//cuttlefish/host/libs/metrics:device_event_type",
-        "//cuttlefish/host/libs/metrics:guest_metrics",
-        "//external_proto:cf_flags_cc_proto",
-        "//external_proto:cf_guest_cc_proto",
-        "//external_proto:cf_metrics_event_v2_cc_proto",
-        "@fmt",
-    ],
-)
-
-cf_cc_library(
     name = "guest_metrics",
     srcs = ["guest_metrics.cc"],
     hdrs = ["guest_metrics.h"],
@@ -151,23 +118,6 @@ cf_cc_library(
         "//cuttlefish/host/libs/metrics:parsed_flags",
         "//cuttlefish/result",
         "@fmt",
-    ],
-)
-
-cf_cc_library(
-    name = "host_conversion",
-    srcs = ["host_conversion.cc"],
-    hdrs = ["host_conversion.h"],
-    deps = [
-        "//cuttlefish/common/libs/utils:host_info",
-        "//cuttlefish/host/libs/metrics:gce_environment",
-        "//cuttlefish/host/libs/metrics:github_environment",
-        "//cuttlefish/host/libs/metrics:host_metrics",
-        "//cuttlefish/host/libs/metrics:invoker",
-        "//external_proto:cf_gce_environment_cc_proto",
-        "//external_proto:cf_github_actions_environment_cc_proto",
-        "//external_proto:cf_host_cc_proto",
-        "//external_proto:cf_metrics_event_v2_cc_proto",
     ],
 )
 
@@ -192,23 +142,6 @@ cf_cc_library(
 )
 
 cf_cc_library(
-    name = "metrics_conversion",
-    srcs = ["metrics_conversion.cc"],
-    hdrs = ["metrics_conversion.h"],
-    deps = [
-        "//cuttlefish/host/libs/metrics:fetch_conversion",
-        "//cuttlefish/host/libs/metrics:fetch_metrics",
-        "//cuttlefish/host/libs/metrics:guest_conversion",
-        "//cuttlefish/host/libs/metrics:guest_metrics",
-        "//cuttlefish/host/libs/metrics:host_conversion",
-        "//cuttlefish/host/libs/metrics:host_metrics",
-        "//external_proto:cf_log_cc_proto",
-        "//external_proto:cf_metrics_event_v2_cc_proto",
-        "@protobuf//:timestamp_cc_proto",
-    ],
-)
-
-cf_cc_library(
     name = "metrics_environment",
     srcs = ["metrics_environment.cc"],
     hdrs = ["metrics_environment.h"],
@@ -225,7 +158,7 @@ cf_cc_library(
         "//cuttlefish/host/libs/log_names",
         "//cuttlefish/host/libs/metrics:enabled",
         "//cuttlefish/host/libs/metrics:host_metrics",
-        "//cuttlefish/host/libs/metrics:metrics_conversion",
+        "//cuttlefish/host/libs/metrics/conversion:metrics_data",
         "//cuttlefish/host/libs/metrics:metrics_transmitter",
         "//cuttlefish/host/libs/metrics:metrics_writer",
         "//cuttlefish/host/libs/metrics:session_id",
