+++ base/cvd/cuttlefish/host/libs/metrics/device_metrics_orchestration.cc
@@ -27,9 +27,9 @@
 #include "cuttlefish/files/file_exists.h"
 #include "cuttlefish/host/commands/cvd/instances/local_instance.h"
 #include "cuttlefish/host/commands/cvd/instances/local_instance_group.h"
+#include "cuttlefish/host/libs/metrics/conversion/metrics_data.h"
 #include "cuttlefish/host/libs/metrics/device_event_type.h"
 #include "cuttlefish/host/libs/metrics/guest_metrics.h"
-#include "cuttlefish/host/libs/metrics/metrics_conversion.h"
 #include "cuttlefish/host/libs/metrics/metrics_orchestration.h"
 #include "cuttlefish/host/libs/metrics/notification.h"
 #include "cuttlefish/host/libs/metrics/parsed_flags.h"
