+++ base/cvd/cuttlefish/host/libs/metrics/fetch_metrics_orchestration.cc
@@ -27,8 +27,8 @@
 #include "cuttlefish/files/file_exists.h"
 #include "cuttlefish/host/commands/cvd/fetch/fetch_cvd.h"
 #include "cuttlefish/host/commands/cvd/fetch/fetch_cvd_parser.h"
+#include "cuttlefish/host/libs/metrics/conversion/metrics_data.h"
 #include "cuttlefish/host/libs/metrics/fetch_metrics.h"
-#include "cuttlefish/host/libs/metrics/metrics_conversion.h"
 #include "cuttlefish/host/libs/metrics/metrics_orchestration.h"
 #include "cuttlefish/host/libs/metrics/notification.h"
 #include "cuttlefish/result/result.h"
